all is good

+++ main.h
@@ -7,10 +7,9 @@
 #include <limits.h>
 #include <stdio.h>
 
-
 int _printf(const char *format, ...);
 void ft_put_str(char *str, int *len);
 void ft_printchar(char c, int *len);
-void _print_number(int nbr, int base, int *len);
+void _print_number(long nbr, int base, int *len);
 
 #endif
